Removed concept of tags. Tests pass.
Some tests were updated. t-negt test now does not result in an empty
fdn. The idea is that beliefs will be retained, even explained, if there
is the ability to do so (on another path), even if the original reason
for the belief was that it was implied by something else that has since
been contracted.